Pull pour màj des changements

+++ src/main/java/fr/eni/projet/encheres/dal/AdresseDAO.java
@@ -8,7 +8,7 @@ public interface AdresseDAO {
 	
 	Adresse create(Adresse adresse); // Create
 	
-	Adresse findById(int noAdresse); // Read
+	Adresse findById(long noAdresse); // Read
 
 	Adresse update(Adresse Adresse); // Update
